Add files via upload
Project Overview::

The Movie Search and Explore Web Application allows users to search for their favorite movies and explore the details of each movie. The application features a user-friendly interface with a search bar, where users can input the name of the movie they're looking for. The application then fetches the movie data from an API and displays the results in a visually appealing format.

Key Features::

Search Functionality: Users can search for movies by typing the name of the movie in the search bar.
Movie Details: The application displays the details of each movie, including the title, poster, release date, and rating.
Responsive Design: The application is designed to be responsive, ensuring a seamless user experience across various devices and screen sizes.

Technical Skills::

HTML: Used to create the structure and content of the web application.
CSS: Used to style the application and make it visually appealing.
JavaScript: Used to add interactivity to the application, including the search functionality and API data fetching.
API Integration: Integrated a movie API to fetch movie data and display it in the application.


Learning Experience::

Throughout this project, I learned many valuable skills, including:

How to design and build a responsive web application.
How to integrate APIs to fetch data and display it in the application.
How to use JavaScript to add interactivity to the application.
How to troubleshoot and debug issues in the application.